% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscistree.R
\name{get_mut_graph}
\alias{get_mut_graph}
\title{Convert a single-cell phylogeny with mutation placements into a mutation graph}
\usage{
get_mut_graph(gtree)
}
\arguments{
\item{gtree}{tbl_graph The single-cell phylogeny}

\item{mut_nodes}{dataframe Mutation placements}
}
\value{
igraph Mutation graph
}
\description{
Convert a single-cell phylogeny with mutation placements into a mutation graph
}
