% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtree_small}
\alias{gtree_small}
\title{Smaller example annotated tree built from P_small}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 199.
}
\usage{
gtree_small
}
\description{
Smaller example annotated tree built from P_small
}
\keyword{datasets}
