% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scistreer.R
\name{ladderize}
\alias{ladderize}
\title{From ape; will remove once new ape version is released
https://github.com/emmanuelparadis/ape/issues/54}
\usage{
ladderize(phy, right = TRUE)
}
\arguments{
\item{phy}{phylo The phylogeny}

\item{right}{logical Whether ladderize to the right}
}
\description{
From ape; will remove once new ape version is released
https://github.com/emmanuelparadis/ape/issues/54
}
\examples{
tree_small = ladderize(tree_small)
}
