% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_nodes_small}
\alias{mut_nodes_small}
\title{Mutation placements calculated from tree_small and P_small}
\format{
An object of class \code{data.frame} with 9 rows and 2 columns.
}
\usage{
mut_nodes_small
}
\description{
Mutation placements calculated from tree_small and P_small
}
\keyword{datasets}
