% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scistreer.R
\name{mut_to_tree}
\alias{mut_to_tree}
\title{Transfer mutation assignment onto a single-cell phylogeny}
\usage{
mut_to_tree(gtree, mut_nodes)
}
\arguments{
\item{gtree}{tbl_graph The single-cell phylogeny}

\item{mut_nodes}{dataframe Mutation placements}
}
\value{
tbl_graph A single-cell phylogeny with mutation placements
}
\description{
Transfer mutation assignment onto a single-cell phylogeny
}
\examples{
gtree_small = mut_to_tree(gtree_small, mut_nodes_small)
}
