% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P_example}
\alias{P_example}
\title{example genotype probability matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 25 columns.
}
\usage{
P_example
}
\description{
example genotype probability matrix
}
\keyword{datasets}
