% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P_small}
\alias{P_small}
\title{Smaller example genotype probability matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 25 columns.
}
\usage{
P_small
}
\description{
Smaller example genotype probability matrix
}
\keyword{datasets}
