% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_phylo_heatmap}
\alias{plot_phylo_heatmap}
\title{Plot phylogeny and mutation heatmap}
\usage{
plot_phylo_heatmap(tree, P, branch_width = 0.5, root_edge = TRUE)
}
\arguments{
\item{tree}{phylo or tbl_graph Phylogeny}

\item{branch_width}{numeric Branch width}

\item{root_edge}{logical Whether to plot root edge}

\item{init}{P matrix Genotype probability matrix}
}
\description{
Plot phylogeny and mutation heatmap
}
