% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scistreer.R
\name{score_tree}
\alias{score_tree}
\title{Score a tree based on maximum likelihood}
\usage{
score_tree(tree, P, get_l_matrix = FALSE)
}
\arguments{
\item{tree}{phylo object}

\item{P}{genotype probability matrix}

\item{get_l_matrix}{whether to compute the whole likelihood matrix}
}
\value{
list Likelihood scores of a tree
}
\description{
Score a tree based on maximum likelihood
}
\examples{
tree_likelihood = score_tree(tree_upgma, P_small)$l_tree
}
