% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_small}
\alias{tree_small}
\title{Smaller example tree built from P_small}
\format{
An object of class \code{phylo} of length 5.
}
\usage{
tree_small
}
\description{
Smaller example tree built from P_small
}
\keyword{datasets}
