% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_upgma}
\alias{tree_upgma}
\title{Example tree built using UPGMA from P_small}
\format{
An object of class \code{phylo} of length 4.
}
\usage{
tree_upgma
}
\description{
Example tree built using UPGMA from P_small
}
\keyword{datasets}
