% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scistreer.R
\name{annotate_tree}
\alias{annotate_tree}
\title{Find maximum lilkelihood assignment of mutations on a tree}
\usage{
annotate_tree(tree, P)
}
\arguments{
\item{tree}{phylo Single-cell phylogenetic tree}

\item{P}{matrix Genotype probability matrix}
}
\value{
tbl_graph A single-cell phylogeny with mutation placements
}
\description{
Find maximum lilkelihood assignment of mutations on a tree
}
\examples{
gtree_small = annotate_tree(tree_small, P_small)
}
