% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plot_phylo_heatmap}
\alias{plot_phylo_heatmap}
\title{Plot phylogeny and mutation heatmap}
\usage{
plot_phylo_heatmap(tree, P, branch_width = 0.5, root_edge = TRUE)
}
\arguments{
\item{tree}{phylo or tbl_graph Phylogeny}

\item{P}{matrix Genotype probability matrix}

\item{branch_width}{numeric Branch width}

\item{root_edge}{logical Whether to plot root edge}
}
\value{
ggplot Plot visualizing the single-cell phylogeny and mutation probability heatmap
}
\description{
Plot phylogeny and mutation heatmap
}
\examples{
p = plot_phylo_heatmap(tree_small, P_small)
}
